DR/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Google Gemini API
google-generativeai==0.3.1

# Vector search and embeddings (optional for advanced RAG)
# sentence-transformers==2.2.2
# chromadb==0.4.18
# faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
